BarcodeMaster/requirements.txt
Pillow
psutil
requests
pyserial
Flask>=2.0.0,<3.0.0
waitress>=2.0.0,<3.0.0
pandas
openpyxl
pyodbc==4.0.35
pyinstaller==5.6.2
